refactor: rename and componentize map-title.tsx
